+++ Libplanet/Tx/TxPolicyViolationException.cs
@@ -11,7 +11,9 @@ namespace Libplanet.Tx
     public class TxPolicyViolationException : InvalidTxException
     {
         /// <summary>
-        /// Creates a new <see cref="TxPolicyViolationException"/> instance.
+        /// Creates a new <see cref="TxPolicyViolationException"/> instance. Use
+        /// <see cref="TxPolicyViolationException(TxId, string, Exception)"/> if innerException
+        /// is required.
         /// </summary>
         /// <param name="txid">The <see cref="Transaction{T}.Id"/> of the
         /// <see cref="Transaction{T}"/> violating a <see cref="IBlockPolicy{T}"/>.
@@ -21,6 +23,22 @@ namespace Libplanet.Tx
         /// </param>
         public TxPolicyViolationException(TxId txid, string message)
             : base(txid, message)
+        {
+        }
+
+        /// <inheritdoc cref="TxPolicyViolationException(TxId, string)"/>
+        /// <param name="txid">The <see cref="Transaction{T}.Id"/> of the
+        /// <see cref="Transaction{T}"/> violating a <see cref="IBlockPolicy{T}"/>.
+        /// Gets prepended to <paramref name="message"/>.</param>
+        /// <param name="message">A description for the reason of violation
+        /// given by an implementation of <see cref="IBlockPolicy{T}"/>. </param>
+        /// <param name="innerException">The <see cref="Exception"/> for
+        /// <see cref="Exception.InnerException"/>.</param>
+        public TxPolicyViolationException(
+            TxId txid,
+            string message,
+            Exception innerException)
+            : base(txid, message, innerException)
         {
         }
     }